Replace Move Actions With Generic Move Command
We didn't need a specific command for each movement when
most of them more or less do the same thing.

With that said we may still require an identity tag when sending the
messages to the client to determine contextual animations.

Along with this, we may also want to wrap the common implementations inside of
skills to reduce mess.