Add Area Of Effect Condition
Some cards may activate in unusual places so we
need to check this to make sure if it is valid.

We may run into a problem when we introduce the concept
of changing control so we could possibly expand this into its
own switch statement or something to make sure to avoid anything
being too coupled.