Split Card Into Model & View (!!!)
Like with the game objects itself we seperate the construction
script from the scene. This lets us have more simplified control over
the data. We replaced the methods roughly though so we'll need to take
another look over them.

We also need to change our command move objects to use rect_global_position
instead of RectGlobalPosition. This seems to be some sort of individual control thing
vs a scene (which understands RGP but base controls do not?)