Add Card Group & Input To The Tree
We'll probably have to create a dedicated card manager object
at some point. Right now we add them all to the card group (if they
weren't in) and then check the mousePos against their position. If absolutely
no cards were selected and the player was attacking, we then deslect and cancel
the attack.

We also add the ability for selection/battle icons to go away post-battle.