Remove Defense Value. Rename Attack To Power.
Our CounterAttack System wasn't making a lot of visual sense.
We believe it may be better to just have a single central power value
if we weren't going to use anything like a yugioh-esque defense mode
to captalize on the defense values existence. We may lose out
some identity but we believe that this is more functionally sublime.

(We hope anyway)