Split Messenger Into Sender And Receiver
The Sender listens for signals and then tells the messenger to
call them. While the Messenger receives queue events and passes
them onto the receiver which then delegates them to the game. This
indirection allows us to seperateo our concerns a bit (and essentially
a hide a laundy list of signals). Not perfect but probably better than what we had.