Replace Signals With Unpack Method
Our previous implementation had too many signals which wasn't nice.

This new implementation uses two new signals. The "Queue" signal
we get from the Server and the Unpack Signal.

When the Queue Signal is called we use the command name (previously
a signal but now just a string) to switch expression out the correct method.

Once we get it, we then give it a one shot connection to our unpack signal.

This connection means that EmitSignal(unpack, values) which unpack the param
object[] array into the method correctly (if the parameters line up correctly)