Inline Commands. Fixes Slow State.
We were clogging our command queue with invisible commands
that didn't need to be queued which led to our state being dark for too
long. Instead we've deleted and inlined them directly on game.

On top of this, instead of await await await, we return a signalawaiter
to the top and await that. (Also it seems like awaiting a tween
signalawaiter triggers it?)